Force int at the description level

+++ homeassistant/components/tesla_fleet/number.py
@@ -33,7 +33,7 @@ PARALLEL_UPDATES = 0
 class TeslaFleetNumberVehicleEntityDescription(NumberEntityDescription):
     """Describes TeslaFleet Number entity."""
 
-    func: Callable[[VehicleFleet, float], Awaitable[Any]]
+    func: Callable[[VehicleFleet, int], Awaitable[Any]]
     native_min_value: float
     native_max_value: float
     min_key: str | None = None
@@ -51,7 +51,7 @@ VEHICLE_DESCRIPTIONS: tuple[TeslaFleetNumberVehicleEntityDescription, ...] = (
         device_class=NumberDeviceClass.CURRENT,
         mode=NumberMode.AUTO,
         max_key="charge_state_charge_current_request_max",
-        func=lambda api, value: api.set_charging_amps(int(value)),
+        func=lambda api, value: api.set_charging_amps(value),
         scopes=[Scope.VEHICLE_CHARGING_CMDS],
     ),
     TeslaFleetNumberVehicleEntityDescription(
@@ -64,7 +64,7 @@ VEHICLE_DESCRIPTIONS: tuple[TeslaFleetNumberVehicleEntityDescription, ...] = (
         mode=NumberMode.AUTO,
         min_key="charge_state_charge_limit_soc_min",
         max_key="charge_state_charge_limit_soc_max",
-        func=lambda api, value: api.set_charge_limit(int(value)),
+        func=lambda api, value: api.set_charge_limit(value),
         scopes=[Scope.VEHICLE_CHARGING_CMDS, Scope.VEHICLE_CMDS],
     ),
 )
@@ -74,19 +74,19 @@ VEHICLE_DESCRIPTIONS: tuple[TeslaFleetNumberVehicleEntityDescription, ...] = (
 class TeslaFleetNumberBatteryEntityDescription(NumberEntityDescription):
     """Describes TeslaFleet Number entity."""
 
-    func: Callable[[EnergySite, float], Awaitable[Any]]
+    func: Callable[[EnergySite, int], Awaitable[Any]]
     requires: str | None = None
 
 
 ENERGY_INFO_DESCRIPTIONS: tuple[TeslaFleetNumberBatteryEntityDescription, ...] = (
     TeslaFleetNumberBatteryEntityDescription(
         key="backup_reserve_percent",
-        func=lambda api, value: api.backup(int(value)),
+        func=lambda api, value: api.backup(value),
         requires="components_battery",
     ),
     TeslaFleetNumberBatteryEntityDescription(
         key="off_grid_vehicle_charging_reserve_percent",
-        func=lambda api, value: api.off_grid_vehicle_charging_reserve(int(value)),
+        func=lambda api, value: api.off_grid_vehicle_charging_reserve(value),
         requires="components_off_grid_vehicle_charging_reserve_supported",
     ),
 )
